backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=10.3.0
python-decouple==3.8
django-filter==23.5
pandas>=2.1.4
numpy>=1.26.3
scikit-learn>=1.4.0
requests==2.31.0
qrcode==7.4.2
gunicorn==21.2.0
whitenoise==6.6.0
mssql-django==1.4
pyodbc==5.0.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
bakong-khqr==0.4.19
